Initial commit with WIP code to gather workouts
Code currently has the ability to log into an account, navigate to the
desired user page, and generate HTML pages for a month's worth of
workouts at a time. Need to finalize the parsing of this HTML into a
more usable format of workout attributes and statistics.

After doing the above, will have to work on GPX export to gather
location data.