Complete workout gathering implementation
Workouts now can be successfully scraped for a given athlete. Two total
web page requests are sent per month of athlete data, resulting in a
relatviely fast gathering speed, as well as staying within the
reasonable use of web requests of a real human user. There are wait
calls to ensure this.

As it stands right now, a single call to this function will generate a
JSON file in the same directory it was run in.

Note, we will not be including the actual data scraped using this
program in the repo in order to comply with Strava's Terms of Service.